backend/requirements.txt
flask 
google-generativeai 
gTTS 
pillow
flask_cors
pygame